requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
fastapi
uvicorn
pillow
python-multipart
mediapipe